refactor: use IssueStatusId type, move Issue to issueTypes.ts
