+++ app/comments/commentTypes.ts
@@ -11,19 +11,6 @@ export interface Comment {
 	position: string | null;
 }
 
-export interface Issue {
-	id: number;
-	status: 'open' | 'closed';
-	created: string;
-	user_id: number;
-	user_name: string | null;
-	user_picture: string | null;
-	user_rating: number | null;
-	user_external_id: string | null;
-	isAdmin: boolean;
-	comments: Comment[];
-}
-
 export interface CommentCounts {
 	total: number;
 	comments: number;
